adjust student

+++ service/student.js
@@ -29,6 +29,7 @@ module.exports = {
       throw err;
     }
   },
+
   // 取得學生詳情資訊
   getStudentDetail: async (req, callback) => {
     try {
@@ -46,16 +47,13 @@ module.exports = {
 
       // student detail
       SQLStr =
-        "SELECT  \
-          s.student_name, \
+        "SELECT s.student_name, \
           s.student_no, \
           s.class_name, \
-          s.total_time - sum(w.end_time - w.start_time) as remaining_time \
-        FROM student s, worktime w \
-        WHERE s.student_no = w.student_no \
-          AND s.semester = w.semester \
-          AND s.semester = ? \
-          AND s.student_no = ?";
+          s.total_time - COALESCE(sum(w.end_time - w.start_time), 0) as remaining_time \
+        FROM student s \
+        LEFT OUTER JOIN worktime w ON w.student_no = s.student_no AND w.semester = s.semester \
+        WHERE s.semester = ? AND s.student_no = ?";
       SQLData = await Promise.resolve(sqlInfo.SQLQuery(conn, SQLStr, [getSemester(), student_no]));
 
       output(
@@ -73,14 +71,15 @@ module.exports = {
       throw err;
     }
   },
+
   // 新增學生
   newStudent: async (req, callback) => {
     try {
-      const { query } = req;
-      const { student_name = "", class_name = "", student_no = "", semester = "" } = query;
+      const { body } = req;
+      const { student_name = "", class_name = "", student_no = "", semester = "" } = body;
 
       // 計算總時數
-      let { total_h, total_m } = query;
+      let { total_h, total_m } = body;
       total_h = parseInt(total_h) || 0;
       total_m = parseInt(total_m) || 0;
       const total_time = total_h * 60 + total_m || 0;
@@ -123,6 +122,7 @@ module.exports = {
       throw err;
     }
   },
+
   // 取得學生列表
   getStudentList: async (req, callback) => {
     try {
@@ -139,13 +139,12 @@ module.exports = {
           s.student_no, \
           s.class_name, \
           s.student_name, \
-          sum(w.end_time - w.start_time) as working_minutes, \
-          s.total_time - sum(w.end_time - w.start_time) as remaining_minutes, \
+          COALESCE(sum(w.end_time - w.start_time), 0) as working_minutes, \
+          s.total_time - COALESCE(sum(w.end_time - w.start_time), 0) as remaining_minutes, \
           s.total_time  as total_minutes \
-        FROM student s, worktime w \
-        WHERE s.student_no = w.student_no \
-          AND s.semester = w.semester \
-          AND s.semester = ? \
+        FROM student s \
+        LEFT OUTER JOIN worktime w ON w.student_no = s.student_no AND w.semester = s.semester \
+        WHERE s.semester = ? \
         GROUP BY student_no \
         ORDER BY student_no ASC";
       SQLData = await Promise.resolve(sqlInfo.SQLQuery(conn, SQLStr, [semester]));
@@ -166,14 +165,16 @@ module.exports = {
       throw err;
     }
   },
+
   // 修改學生資料
   updateStudent: async (req, callback) => {
     try {
-      const { query } = req;
-      const { id = 0, student_name = "", class_name = "", student_no = "", semester = "" } = query;
+      const { query, body } = req;
+      const { id = 0 } = query;
+      const { student_name = "", class_name = "", student_no = "", semester = "" } = body;
 
       // 計算總時數
-      let { total_h, total_m } = query;
+      let { total_h, total_m } = body;
       total_h = parseInt(total_h) || 0;
       total_m = parseInt(total_m) || 0;
       const total_time = total_h * 60 + total_m || 0;
@@ -201,7 +202,7 @@ module.exports = {
       };
 
       // 更新學生資料
-      SQLStr = "UPDATE member SET ? WHERE id = ?";
+      SQLStr = "UPDATE student SET ? WHERE id = ?";
       SQLFlag = await Promise.resolve(sqlInfo.SQLQuery(conn, SQLStr, [updateQuery, id]));
       if (!SQLFlag) {
         output(callback, { code: 403 }, { conn });
@@ -214,6 +215,7 @@ module.exports = {
       throw err;
     }
   },
+
   // 刪除學生資料
   deleteStudent: async (req, callback) => {
     try {
@@ -228,6 +230,12 @@ module.exports = {
       // connection SQL
       const conn = sqlInfo.conn("worktime");
       let [SQLStr, SQLData, SQLFlag] = ["", [], undefined];
+      SQLStr = "SELECT count(*) as size FROM student WHERE semester=? AND student_no=?";
+      SQLData = await Promise.resolve(sqlInfo.SQLQuery(conn, SQLStr, [semester, student_no]));
+      if (SQLData[0]["size"] !== 1) {
+        output(callback, { code: 410 }, { conn });
+        return;
+      }
 
       // 刪除學生的時數
       SQLStr = "DELETE FROM worktime WHERE semester=? AND student_no=?";
@@ -251,11 +259,12 @@ module.exports = {
       throw err;
     }
   },
+
   // 取得前學期學生列表
   getOldStudentList: async (req, callback) => {
     try {
       const { query } = req;
-      let { semester = "" } = query;
+      const { semester = "" } = query;
 
       if (semester === "") {
         output(callback, { code: 400 }, { conn });
@@ -264,24 +273,38 @@ module.exports = {
 
       // 取得前學期
       const semester_str = semester.split("-");
-      semester = semester_str[1] === 1 ? `${semester_str[0]}-2` : `${semester_str[0] + 1}-1`;
+      const old_semester =
+        parseInt(semester_str[1]) === 1 ? `${parseInt(semester_str[0]) - 1}-2` : `${parseInt(semester_str[0])}-1`;
 
       // connection SQL
       const conn = sqlInfo.conn("worktime");
       let [SQLStr, SQLData] = ["", []];
 
       // 取得前學期之學生列表
-      SQLStr = "SELECT * FROM student WHERE semester=? ORDER BY student_no DESC";
-      SQLData = await Promise.resolve(sqlInfo.SQLQuery(conn, SQLStr, [semester]));
+      SQLStr =
+        "SELECT \
+          s.id, \
+          s.student_no, \
+          s.class_name, \
+          s.student_name, \
+          COALESCE(sum(w.end_time - w.start_time), 0) as working_minutes, \
+          s.total_time - COALESCE(sum(w.end_time - w.start_time), 0) as remaining_minutes, \
+          s.total_time  as total_minutes \
+        FROM \
+          student s \
+        LEFT OUTER JOIN worktime w ON w.student_no = s.student_no AND w.semester = s.semester \
+        WHERE \
+          s.semester=? \
+          AND s.student_no NOT IN (SELECT student_no FROM student WHERE semester=?) \
+        GROUP BY student_no \
+        ORDER BY student_no ASC";
+      SQLData = await Promise.resolve(sqlInfo.SQLQuery(conn, SQLStr, [old_semester, semester]));
 
       output(
         callback,
         {
           code: 200,
-          data: {
-            size: SQLData.length,
-            results: [...SQLData],
-          },
+          data: { semester: old_semester, size: SQLData.length, results: [...SQLData] },
         },
         { conn }
       );
